Time: 42 ms (44.97%), Space: 100.6 MB (35.29%) - LeetHub

+++ 3430-count-days-without-meetings/3430-count-days-without-meetings.java
@@ -5,7 +5,7 @@ class Solution {
         // Sort meetings based on starting times
         Arrays.sort(meetings, Comparator.comparingInt(a -> a[0]));
         int lastDay = 0;
-        long totalMeetings = 0;
+        int totalMeetings = 0;
         for(int i = 0;i<meetings.length;i++){
             int temp      = meetings[i][1] - Math.max(meetings[i][0],lastDay) + 1;
             totalMeetings += temp>0 ? temp: 0;
